Load resources dynamically based on host name mapping
