--- app/config.default.js
@@ -1,453 +1,0 @@
-const CONFIG = process.env.CONFIG || 'default';
-const API_URL = process.env.API_URL || 'https://dev-api.digitransit.fi';
-const MAP_URL = process.env.MAP_URL || 'https://{s}-dev-api.digitransit.fi';
-const APP_PATH = process.env.APP_CONTEXT || '';
-const PIWIK_ADDRESS = process.env.PIWIK_ADDRESS || '';
-const PIWIK_ID = process.env.PIWIK_ID || '';
-const SENTRY_DSN = process.env.SENTRY_DSN || '';
-const PORT = process.env.PORT || 8080;
-const APP_DESCRIPTION = 'Digitransit journey planning UI';
-
-export default {
-  PIWIK_ADDRESS,
-  PIWIK_ID,
-  SENTRY_DSN,
-  PORT,
-  CONFIG,
-
-  URL: {
-    API_URL,
-    OTP: `${API_URL}/routing/v1/routers/finland/`,
-    MAP: `${MAP_URL}/map/v1/hsl-map/`,
-    STOP_MAP: `${API_URL}/map/v1/finland-stop-map/`,
-    CITYBIKE_MAP: `${API_URL}/map/v1/hsl-citybike-map/`,
-    MQTT: 'wss://dev.hsl.fi/mqtt-proxy',
-    ALERTS: `${API_URL}/realtime/service-alerts/v1`,
-    FONT: 'https://fonts.googleapis.com/css?family=Lato:300,400,900%7CPT+Sans+Narrow:400,700',
-    REALTIME: `${API_URL}/realtime/vehicle-positions/v1`,
-    PELIAS: `${API_URL}/geocoding/v1/search`,
-    PELIAS_REVERSE_GEOCODER: `${API_URL}/geocoding/v1/reverse`,
-  },
-
-  APP_PATH: `${APP_PATH}`,
-  title: 'Reittihaku',
-
-  contactName: {
-    sv: 'Digitransit',
-    fi: 'Digitransit',
-    default: "Digitransit's",
-  },
-
-  searchParams: {},
-
-  search: {
-    suggestions: {
-      useTransportIcons: false,
-    },
-  },
-
-  nearbyRoutes: {
-    radius: 10000,
-    bucketSize: 1000,
-  },
-
-  maxWalkDistance: 10000,
-  maxBikingDistance: 100000,
-  availableLanguages: ['fi', 'sv', 'en', 'fr', 'nb', 'de'],
-  defaultLanguage: 'en',
-  // This timezone data will expire on 31.12.2020
-  timezoneData: 'Europe/Helsinki|EET EEST|-20 -30|01010101010101010101010|1BWp0 1qM0 WM0 1qM0 ' +
-    'WM0 1qM0 11A0 1o00 11A0 1o00 11A0 1o00 11A0 1qM0 WM0 1qM0 WM0 1qM0 11A0 1o00 11A0 1o00|35e5',
-
-  mainMenu: {
-    // Whether to show the left menu toggle button at all
-    show: true,
-    showDisruptions: true,
-    showInquiry: false,
-    showLoginCreateAccount: true,
-    showOffCanvasList: true,
-  },
-
-  feedback: {
-    // Whether to allow the feedback popup
-    enable: false,
-  },
-
-  itinerary: {
-    // How long vehicle should be late in order to mark it delayed. Measured in seconds.
-    delayThreshold: 180,
-    // Wait time to show "wait leg"? e.g. 180 means over 3 minutes are shown as wait time.
-    // Measured in seconds.
-    waitThreshold: 180,
-    enableFeedback: false,
-
-    timeNavigation: {
-      enableButtonArrows: false,
-    },
-  },
-
-  initialLocation: {
-    zoom: 11,
-    lat: 60.17332,
-    lon: 24.94102,
-  },
-
-  nearestStopDistance: {
-    maxShownDistance: 5000,
-  },
-
-  map: {
-    useRetinaTiles: true,
-    tileSize: 512,
-    zoomOffset: -1,
-    useVectorTiles: true,
-
-    genericMarker: {
-      // Do not render name markers at zoom levels below this value
-      nameMarkerMinZoom: 18,
-
-      popup: {
-        offset: [106, 16],
-        maxWidth: 250,
-        minWidth: 250,
-      },
-    },
-
-    line: {
-      halo: {
-        weight: 7,
-        thinWeight: 4,
-      },
-
-      leg: {
-        weight: 5,
-        thinWeight: 2,
-      },
-    },
-
-    useModeIconsInNonTileLayer: false,
-  },
-
-  stopCard: {
-    header: {
-      showDescription: true,
-      showStopCode: true,
-      showDistance: true,
-    },
-  },
-
-  autoSuggest: {
-    // Let Pelias suggest based on current user location
-    locationAware: true,
-  },
-
-  // TODO: Switch back in april
-  cityBike: {
-    showCityBikes: false,
-
-    useUrl: {
-      fi: 'https://www.hsl.fi/citybike',
-      sv: 'https://www.hsl.fi/sv/citybike',
-      en: 'https://www.hsl.fi/en/citybike',
-    },
-
-    infoUrl: {
-      fi: 'https://www.hsl.fi/kaupunkipyörät',
-      sv: 'https://www.hsl.fi/sv/stadscyklar',
-      en: 'https://www.hsl.fi/en/citybikes',
-    },
-
-    cityBikeMinZoom: 14,
-    cityBikeSmallIconZoom: 14,
-    // When should bikeshare availability be rendered in orange rather than green
-    fewAvailableCount: 3,
-  },
-  // Lowest level for stops and terminals are rendered
-  stopsMinZoom: 13,
-  // Highest level when stops and terminals are still rendered as small markers
-  stopsSmallMaxZoom: 14,
-  // Highest level when terminals are still rendered instead of individual stops
-  terminalStopsMaxZoom: 17,
-  terminalStopsMinZoom: 12,
-  terminalNamesZoom: 16,
-
-  appBarLink: { name: 'Digitransit', href: 'https://www.digitransit.fi/' },
-
-  colors: {
-    primary: '#00AFFF',
-  },
-
-  disruption: {
-    showInfoButton: true,
-  },
-
-  agency: {
-    show: true,
-  },
-
-  socialMedia: {
-    title: 'Digitransit',
-    description: APP_DESCRIPTION,
-    locale: 'en_US',
-
-    twitter: {
-      site: '@hsldevcom',
-    },
-  },
-
-  meta: {
-    description: APP_DESCRIPTION,
-    keywords: 'digitransit',
-  },
-  // Ticket information feature toggle
-  showTicketInformation: false,
-  showRouteInformation: false,
-
-  modeToOTP: {
-    bus: 'BUS',
-    tram: 'TRAM',
-    rail: 'RAIL',
-    subway: 'SUBWAY',
-    citybike: 'BICYCLE_RENT',
-    airplane: 'AIRPLANE',
-    ferry: 'FERRY',
-    walk: 'WALK',
-    bicycle: 'BICYCLE',
-    car: 'CAR',
-  },
-  // Control what transport modes that should be possible to select in the UI
-  // and whether the transport mode is used in trip planning by default.
-  transportModes: {
-    bus: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-
-    tram: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-
-    rail: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-
-    subway: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-
-    // TODO: Switch back in april
-    citybike: {
-      availableForSelection: false,
-      defaultValue: false,
-    },
-
-    airplane: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-
-    ferry: {
-      availableForSelection: true,
-      defaultValue: true,
-    },
-  },
-
-  streetModes: {
-    walk: {
-      availableForSelection: true,
-      defaultValue: true,
-      icon: 'walk',
-    },
-
-    bicycle: {
-      availableForSelection: true,
-      defaultValue: false,
-      icon: 'bicycle-withoutBox',
-    },
-
-    car: {
-      availableForSelection: true,
-      defaultValue: false,
-      icon: 'car-withoutBox',
-    },
-
-    car_park: {
-      availableForSelection: false,
-      defaultValue: false,
-      icon: 'car_park-withoutBox',
-    },
-  },
-
-  ticketOptions: [{
-    displayName: 'Ei lippuvyöhykerajoitusta',
-    value: '0',
-  }],
-
-  accessibilityOptions: [{
-    displayName: 'Ei rajoitusta',
-    value: '0',
-  }, {
-    displayName: 'Liikun pyörätuolilla',
-    value: '1',
-  }],
-
-  showModeFilter: true,
-
-  moment: {
-    relativeTimeThreshold: {
-      seconds: 55,
-      minutes: 59,
-      hours: 23,
-      days: 26,
-      months: 11,
-    },
-  },
-
-  customizeSearch: {
-    walkReluctance: {
-      available: true,
-    },
-
-    walkBoardCost: {
-      available: true,
-    },
-
-    transferMargin: {
-      available: true,
-    },
-
-    walkingSpeed: {
-      available: true,
-    },
-
-    ticketOptions: {
-      available: true,
-    },
-
-    accessibility: {
-      available: true,
-    },
-  },
-
-  areaPolygon: [
-    [18.776, 60.3316],
-    [18.9625, 60.7385],
-    [19.8615, 60.8957],
-    [20.4145, 61.1942],
-    [20.4349, 61.9592],
-    [19.7853, 63.2157],
-    [20.4727, 63.6319],
-    [21.6353, 63.8559],
-    [23.4626, 64.7794],
-    [23.7244, 65.3008],
-    [23.6873, 65.8569],
-    [23.2069, 66.2701],
-    [23.4627, 66.8344],
-    [22.9291, 67.4662],
-    [23.0459, 67.9229],
-    [20.5459, 68.7605],
-    [20.0996, 69.14],
-    [21.426, 69.4835],
-    [21.9928, 69.4009],
-    [22.9226, 68.8678],
-    [23.8108, 69.0145],
-    [24.6903, 68.8614],
-    [25.2262, 69.0596],
-    [25.4029, 69.7235],
-    [26.066, 70.0559],
-    [28.2123, 70.2496],
-    [29.5813, 69.7854],
-    [29.8467, 69.49],
-    [28.9502, 68.515],
-    [30.4855, 67.6952],
-    [29.4962, 66.9232],
-    [30.5219, 65.8728],
-    [30.1543, 64.9646],
-    [30.9641, 64.1321],
-    [30.572, 63.7098],
-    [31.5491, 63.3309],
-    [31.9773, 62.9304],
-    [31.576, 62.426],
-    [27.739, 60.1117],
-    [26.0945, 59.8015],
-    [22.4235, 59.3342],
-    [20.2983, 59.2763],
-    [19.3719, 59.6858],
-    [18.7454, 60.1305],
-    [18.776, 60.3316],
-  ],
-
-  footer: {
-    content: [
-      { label: (function () { return `© HSL, Liikennevirasto ${(1900 + new Date().getYear())}`; }()) },
-      {},
-      { name: 'footer-feedback', nameEn: 'Send feedback', href: 'https://github.com/HSLdevcom/digitransit-ui/issues', icon: 'icon-icon_speech-bubble' },
-      { name: 'about-this-service', nameEn: 'About this service', route: '/tietoja-palvelusta', icon: 'icon-icon_info' },
-    ],
-  },
-
-  // Default origin endpoint to use when user is outside of area
-  defaultEndpoint: {
-    address: 'Helsinki-Vantaan Lentoasema',
-    lat: 60.317429,
-    lon: 24.9690395,
-  },
-  defaultOrigins: [
-    { icon: 'icon-icon_airplane', label: 'Helsinki-Vantaan lentoasema', lat: 60.317429, lon: 24.9690395 },
-    { icon: 'icon-icon_ferry', label: 'Turun satama', lat: 60.436363, lon: 22.220002 },
-    { icon: 'icon-icon_airplane', label: 'Rovaniemen lentoasema', lat: 66.557326, lon: 25.828135 },
-  ],
-
-  /* eslint-disable max-len*/
-  aboutThisService: {
-    fi: {
-      about: 'Palvelu kattaa joukkoliikenteen, kävelyn, pyöräilyn ja yksityisautoilun rajatuilta osin. Palvelu perustuu Digitransit palvelualustaan.',
-      digitransit: 'Digitransit palvelualusta on HSL:n ja Liikenneviraston kehittämä avoimen lähdekoodin ratkaisu, jonka taustalla toimii mm. OpenTripPlanner. Lähdekoodi tarjotaan EUPL v1.2 ja AGPLv3 lisensseillä.',
-      datasources: 'Kartat, kadut, rakennukset, pysäkkisijainnit ym. tiedot tarjoaa © OpenStreetMap contributors ja ne ladataan Geofabrik palvelusta. Osoitetiedot tuodaan VRK:n rakennustietorekisteristä ja ne ladataan OpenAddresses-palvelusta. Joukkoliikenteen reitit ja aikataulut ladataan mm. Liikenneviraston valtakunnallisesta joukkoliikenteen tietokannasta.',
-    },
-
-    sv: {
-      about: 'Reseplaneraren täcker med vissa begränsningar kollektivtrafik, promenad, cykling samt privatbilism. Tjänsten baserar sig på Digitransit-plattformen.',
-      digitransit: 'Digitransit-plattformen är en öppen programvara utvecklad av HRT och Trafikverket, som bl.a. stödjer sig på OpenTripPlanner. Källkoden distribueras under EUPL v1.2 och AGPLv3 licenserna.',
-      datasources: 'Kartor, gator, byggnader, hållplatser och dylik information erbjuds av © OpenStreetMap contributors och laddas ned från Geofabrik-tjänsten. Addressinformation hämtas från BRC:s byggnadsinformationsregister och laddas ned från OpenAddresses-tjänsten. Kollektivtrafikens rutter och tidtabeller hämtas bl.a. från Trafikverkets landsomfattande kollektivtrafiksdatabas.',
-    },
-
-    en: {
-      about: 'The service covers public transport, walking, cycling, and some private car use. Service is built on Digitransit platform.',
-      digitransit: 'Digitransit service platform is created by HSL and Finnish Transport Agency, built on top of e.g. OpenTripPlanner. The source code of the platform is dual-licensed under the EUPL v1.2 and AGPLv3 licenses.',
-      datasources: "Maps, streets, buildings, stop locations etc. from © OpenStreetMap contributors downloaded from Geofabrik. Additional address data from Finland's Population Register Centre downloaded from OpenAddresses Public transport routes and timetables from Finnish Transport Agency's national public transit database.",
-    },
-
-    nb: {},
-    fr: {},
-    de: {},
-  },
-  /* eslint-enable max-len*/
-
-  /* eslint-disable max-len*/
-  staticMessages: [{
-    id: 1,
-
-    content: {
-      fi: {
-        title: 'Tämä on Digitransitin kehitysversio',
-        content: 'Käytät Digitransitin kehitysversiota.',
-      },
-
-      sv: {
-        title: 'Det här är utvecklingsversionen av Digitransit',
-        content: 'Du använder utvecklingsversionen av Digitransit.',
-      },
-
-      en: {
-        title: 'This version of Digitransit is under development',
-        content: 'You are using the development version of Digitransit.',
-      },
-    },
-  }],
-  /* eslint-enable max-len*/
-
-};
